Finished all the CRUD operations

+++ src/main/java/edu/javeriana/abetbackend/CRUD/Services/Find/OutcomeFinder.java
@@ -1,7 +1,7 @@
 package edu.javeriana.abetbackend.CRUD.Services.Find;
 
 import edu.javeriana.abetbackend.Entities.Outcome;
-import edu.javeriana.abetbackend.Exceptions.OutcomeNotFoundById;
+import edu.javeriana.abetbackend.Exceptions.NotFound.OutcomeNotFoundById;
 import edu.javeriana.abetbackend.Repositories.OutcomeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,7 +15,7 @@ public class OutcomeFinder {
     @Autowired
     private OutcomeRepository repository;
 
-    public Outcome findOutcomeById(Long outcomeId){
+    public Outcome findOutcomeById(Integer outcomeId){
         Optional<Outcome> optionalOutcome = repository.findById(outcomeId);
         if (optionalOutcome.isEmpty())
             throw new OutcomeNotFoundById("The outcome with Id " + outcomeId + " was not found");
