+++ src/main/java/edu/javeriana/abetbackend/Repositories/OutcomeRepository.java
@@ -3,6 +3,6 @@ package edu.javeriana.abetbackend.Repositories;
 import edu.javeriana.abetbackend.Entities.Outcome;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OutcomeRepository extends CrudRepository<Outcome,Long> {
+public interface OutcomeRepository extends CrudRepository<Outcome,Integer> {
 
 }
